refactor backend directories structure
