+++ docker/backend/Dockerfile-backend
@@ -33,6 +33,6 @@ COPY --from=backend-builder ${VIRTUAL_ENV} ${VIRTUAL_ENV}
 COPY --from=backend-builder /root/nltk_data /root/nltk_data
 COPY ./data ./data
 COPY ./ai ./ai
-COPY ./backend/web ./web
+COPY ./backend ./backend
 
-CMD uvicorn web.main:app --host 0.0.0.0 --port 80 --log-level warning
+CMD uvicorn backend.main:app --host 0.0.0.0 --port 80 --log-level warning
