+++ docker/nginx/default.conf
@@ -7,7 +7,7 @@ server {
         resolver 127.0.0.11;
 
         rewrite /api/(.*) /$1 break;
-        proxy_pass http://keytokens-analysis-app:80;
+        proxy_pass http://keytokens-analysis-backend:80;
         proxy_redirect off;
         sub_filter 'href="/' 'href="/api';
         sub_filter '/openapi.json' '/api/openapi.json';
